requirements.txt
Django==4.2.13

mysqlclient==2.2.4

# uWSGI==2.0.26

djangorestframework==3.15.1

djangorestframework-simplejwt==5.3.1

django-environ==0.11.2

PyJWT==2.8.0
